move dialog switcher button
