+++ src/app/shared/components/dialog-switcher-button/dialog-switcher-button.component.spec.ts
